requirements.txt
# Flask Core
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Environment & Config
python-dotenv==1.0.0

# Security & Cryptography
passlib[argon2]==1.7.4
PyJWT==2.8.0
cryptography==41.0.7

# HTTP Client
requests==2.31.0

# Optional but recommended
Flask-CORS==4.0.0
marshmallow==3.20.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
faker==20.1.0

# Production server (opcional)
gunicorn==21.2.0
